requirements.txt
pytest==8.3.3
selenium==4.27.1

python-dotenv~=1.0.1
Faker~=36.2.2